Make it compile

+++ examples/build-tree-no-commit.rs
@@ -105,7 +105,7 @@ fn main() -> Result<(), heed::BoxedError> {
         "That makes a total of: {} tree nodes. {:.2}% of all the nodes",
         split_nodes_sum + descendants_sum,
         (split_nodes_sum + descendants_sum) as f64
-            / (split_nodes_sum + descendants_sum + leaf) as f64
+            / (split_nodes_sum as u64 + descendants_sum as u64 + leaf) as f64
             * 100.,
     );
